+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `CreatesOrUpdatesUser` trait
 - Added newly added permissions in `ApiSeed` command
 - Added `user.unverified_new_only` config value
+- Added user impersonation
 
 ### Changed